+++ changelog.md
@@ -11,6 +11,7 @@ This project attempts to adheres to [Semantic Versioning](http://semver.org/).
 - Iterables.filter returning a new iterable containing only those elements for which the predicate returns true
 - Iterables.flatten transforms a nested collection of iterables into a single iterable
 - Iterables.addAll adds all of the input collection into the passed iterable
+- Iterables.cycle returns an infinite iterable that cycles through the input elements (does not support removing elements)
 - Options.nullSafe transforms a null producing function into one returning an option
 - Options.toOption returns a function that builds an option
 - Option.fromOptional and Option.toOptional for interoperability with java.util.Optional